Added file upload to add modal
