+++ frontend/src/hooks/use-add-job.hook.ts
@@ -1,9 +1,10 @@
 import { JobFormValues } from "@/constants/job-form.constants";
 import { useAddJobMutation, useCreateJobQuery } from "./use-query.hook";
 import { useNavigate, useRouter } from "@tanstack/react-router";
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 
 const useAddJob = (token?: string | null) => {
+  const imageInputRef = useRef<HTMLInputElement | null>(null);
   const [jobImage, setJobImage] = useState<string | undefined>();
   const { data } = useCreateJobQuery();
   const { mutate, isPending } = useAddJobMutation();
@@ -34,7 +35,38 @@ const useAddJob = (token?: string | null) => {
     router.history.back();
   };
 
-  return { onSubmit, onClose, setJobImage, jobImage, isPending };
+  const addFile = () => {
+    if (!imageInputRef.current) {
+      throw new Error("Image input ref is not set");
+    }
+
+    imageInputRef.current.onchange = (e) => {
+      const file = (e.target as HTMLInputElement).files?.[0];
+      if (!file) {
+        return;
+      }
+      const reader = new FileReader();
+      reader.onload = async (e) => {
+        const image = e.target?.result as string;
+        if (image) {
+          setJobImage(image);
+        }
+      };
+      reader.readAsDataURL(file);
+    };
+
+    imageInputRef.current.click();
+  };
+
+  return {
+    onSubmit,
+    onClose,
+    setJobImage,
+    addFile,
+    jobImage,
+    isPending,
+    imageInputRef,
+  };
 };
 
 export default useAddJob;
